CRNN/requirements.txt
comtypes
pdf2image
numpy
sklearn
pickle
matplotlib
python-docx
keras